rename: DiagnosisQuizDto 경로 수정
