+++ src/main/java/com/ssafy/soljigi/diagnosis/service/LanguageService.java
@@ -7,7 +7,7 @@ import java.util.Random;
 
 import org.springframework.stereotype.Service;
 
-import com.ssafy.soljigi.diagnosis.dto.DiagnosisQuizDto;
+import com.ssafy.soljigi.diagnosis.dto.response.DiagnosisQuizDto;
 import com.ssafy.soljigi.diagnosis.entity.DiagnosisType;
 
 import lombok.RequiredArgsConstructor;
